增加es查询构造器，修改es查询方法

+++ browser-common/src/test/java/com/platon/browser/utils/EpochUtilTest.java
@@ -13,25 +13,25 @@ public class EpochUtilTest {
 
 	@Test
 	public void testGetEpoch() throws BlockNumberException {
-		assertTrue((EpochUtil.getEpoch(BigInteger.valueOf(1), 5l)==1));
+		assertTrue((EpochUtil.getEpoch(BigInteger.valueOf(1), BigInteger.valueOf(5)).intValue()==1));
 	}
 
 	@Test
 	public void testGetPreEpochLastBlockNumber() throws BlockNumberException {
-		assertEquals(0,EpochUtil.getPreEpochLastBlockNumber(BigInteger.valueOf(1),5L).longValue());
-		assertEquals(0,EpochUtil.getPreEpochLastBlockNumber(BigInteger.valueOf(5),5L).longValue());
-		assertEquals(5,EpochUtil.getPreEpochLastBlockNumber(BigInteger.valueOf(6),5L).longValue());
-		assertEquals(5,EpochUtil.getPreEpochLastBlockNumber(BigInteger.valueOf(10),5L).longValue());
-		assertEquals(10,EpochUtil.getPreEpochLastBlockNumber(BigInteger.valueOf(11),5L).longValue());
+		assertEquals(0,EpochUtil.getPreEpochLastBlockNumber(BigInteger.valueOf(1),BigInteger.valueOf(5)).longValue());
+		assertEquals(0,EpochUtil.getPreEpochLastBlockNumber(BigInteger.valueOf(5),BigInteger.valueOf(5)).longValue());
+		assertEquals(5,EpochUtil.getPreEpochLastBlockNumber(BigInteger.valueOf(6),BigInteger.valueOf(5)).longValue());
+		assertEquals(5,EpochUtil.getPreEpochLastBlockNumber(BigInteger.valueOf(10),BigInteger.valueOf(5)).longValue());
+		assertEquals(10,EpochUtil.getPreEpochLastBlockNumber(BigInteger.valueOf(11),BigInteger.valueOf(5)).longValue());
 	}
 
 	@Test
 	public void testGetCurEpochLastBlockNumber() throws BlockNumberException {
-		assertEquals(5,EpochUtil.getCurEpochLastBlockNumber(BigInteger.valueOf(1),5L).longValue());
-		assertEquals(5,EpochUtil.getCurEpochLastBlockNumber(BigInteger.valueOf(5),5L).longValue());
-		assertEquals(10,EpochUtil.getCurEpochLastBlockNumber(BigInteger.valueOf(6),5L).longValue());
-		assertEquals(10,EpochUtil.getCurEpochLastBlockNumber(BigInteger.valueOf(10),5L).longValue());
-		assertEquals(15,EpochUtil.getCurEpochLastBlockNumber(BigInteger.valueOf(11),5L).longValue());
+		assertEquals(5,EpochUtil.getCurEpochLastBlockNumber(BigInteger.valueOf(1),BigInteger.valueOf(5)).longValue());
+		assertEquals(5,EpochUtil.getCurEpochLastBlockNumber(BigInteger.valueOf(5),BigInteger.valueOf(5)).longValue());
+		assertEquals(10,EpochUtil.getCurEpochLastBlockNumber(BigInteger.valueOf(6),BigInteger.valueOf(5)).longValue());
+		assertEquals(10,EpochUtil.getCurEpochLastBlockNumber(BigInteger.valueOf(10),BigInteger.valueOf(5)).longValue());
+		assertEquals(15,EpochUtil.getCurEpochLastBlockNumber(BigInteger.valueOf(11),BigInteger.valueOf(5)).longValue());
 	}
 
 }
